move web maplibre out of common packs
