+++ apps/app-web/src/machine/map-tools/MapTools.tsx
@@ -21,10 +21,52 @@ export const MapTools: FC<Props> = ({ map, children }) => {
     const [containerRef, setContainerRef] = useState<HTMLElement | null>(null);
     const [cssLoaded, setCssLoaded] = useState(false);
     const [isInitialised, setIsInitialised] = useSubjectState(cartomancer.isInitialised$);
-    const [isStyleLoaded] = useSubjectState(cartomancer.isStyleLoaded$);
+    const [isStyleLoaded, setIsStyleLoaded] = useSubjectState(cartomancer.isStyleLoaded$);
     const [selectedStyle] = useSubjectState(cartomancer.selectedStyle$);
     const [_mapZoom, setMapZoom] = useSubjectState(cartomancer.zoom$);
     const [_mapBearing, setMapBearing] = useSubjectState(cartomancer.bearing$);
+
+    /**
+     * Safely updates style and resolves when the `map.isStyleLoaded()` check resolves.
+     */
+    const updateStyle = async (
+        map: maplibregl.Map,
+        style: string | maplibregl.StyleSpecification,
+        abortSignal: AbortSignal,
+        onError?: (err: unknown) => void
+    ) => {
+        try {
+            setIsStyleLoaded(false);
+            map.setStyle(style);
+            await validateStyleLoaded(map, abortSignal);
+            setIsStyleLoaded(true);
+        } catch (err) {
+            onError?.(err);
+        }
+    };
+
+    /**
+     * Subscribes to map `idle` events and resolves when `map.isStyleLoaded()` resolves.
+     */
+    const validateStyleLoaded = (
+        map: maplibregl.Map,
+        abortSignal: AbortSignal
+    ): Promise<void> => {
+        return new Promise((resolve, reject) => {
+            const isLoadedHandler = (_event: maplibregl.MapDataEvent) => {
+                if (abortSignal.aborted) {
+                    map.off('idle', isLoadedHandler);
+                    reject("User aborted map style validation.")
+                } else
+                    if (map.isStyleLoaded()) {
+                        map.off('idle', isLoadedHandler);
+                        resolve();
+                    }
+            }
+
+            map.on('idle', isLoadedHandler);
+        });
+    };
 
     useEffect(() => {
         const abortController = new AbortController();
@@ -128,7 +170,7 @@ export const MapTools: FC<Props> = ({ map, children }) => {
         }
 
         const abortController = new AbortController();
-        cartomancer.updateStyle(map, nextStyle.style, abortController.signal, (err) => {
+        updateStyle(map, nextStyle.style, abortController.signal, (err) => {
             if (!abortController.signal.aborted) {
                 console.error(err)
             }
