+++ apps/packages/apparatus/src/state-warden/cartomancer/hooks/index.ts
@@ -1,2 +1,1 @@
-export * from './useMapLayerData';
 export * from './useUpdateSourceData';
