+++ apps/ui/web/src/index.ts
@@ -1,2 +1,3 @@
 export * from './forms';
+export * from './map';
 export * as Icons from './icons';
